Update Arbidex URL in SuccessZapModal component
Changed the URL for Arbidex platform in the SuccessZapModal component. The old URL "https://arbidex.fi/farms" was replaced with the new one "https://arbidex.fi/xarxfarms". This was necessary as the old link was broken or outdated. The change will direct users to the correct page when they click on the link.