Add loading state to swap module
This revision introduces a loading state to the 'swap-module' and 'SwapForm.vue'. This provides feedback to users, via UI changes, when prices of tokens are being fetched or updated. The 'input-style' element now shows a progress spinner while the token prices are being loaded, and the regular content is hidden. Similarly, the approximation of token price in 'usd-equal-text' is hidden while prices are loading, and replaced with a simple '-'. In 'SwapForm.vue', token amounts are updated based on the fetched token prices.