Add tracking actions to multiple components
The changes aim at improving action tracking across the application. Included in several VUE components are new mapActions that call the `trackClick` action from the track module. This will enable tracking for specific user interactions like clicking on the swap-in/swap-out on the featured page, clicking on the swap option in the mobile and desktop menu, form swap clicks, and page view events. This will aid in user-behavior analysis for enhanced user interface/experience decisions.