Add zkSync fee refund notice and clear fee history after modal close
A new information section has been added to the Redeem.vue file to inform users that 30-90% of the displayed gas fee on zkSync will be refunded automatically. This notice is only shown if the network name is 'zkSync'. Some new CSS styling was included to accommodate the layout of this new notice.

Additionally, to ensure proper state management, the zkSync fee history is now cleared every time the success modal is closed. This change was applied to both odos-swap.js and success-modal.js files. These changes together aim to improve user understanding of transaction fees and keep application state consistent.