Add gas fee info display for zkSync network
Added a new interface block to the SwapForm component. In this block, when the connected network is zkSync, a message display indicates that 30-90% of the displayed gas fee will be refunded automatically. This change is necessary to inform users about the beneficial feature of zkSync. The relevant styling and getter methods were also added to support this change.