Add SlippageError component and update styling for slippage settings
Added a new SlippageError component to handle slippage-related errors. This includes creating a new Vue file with the required markup, script and styles, as well as updating ErrorModal.vue to support the new error type. The slippage settings UI has been updated to include a tooltip with an explanation of the different slippage settings and a warning for potential transaction failures when a low slippage setting is used. Also, error modal has been updated to handle the slippage errors.