Refactor slippage tolerance information layout
This commit updates the layout of the slippage tolerance information in both `SlippageError.vue` and `SwapSlippageSettings.vue` components. This modification made the information clearer by displaying each piece of data individually instead of one block of text. The use of tooltips in the `SwapSlippageSettings.vue` component was implemented to present these data in a more user-friendly manner. Furthermore, the descriptions for each level of slippage settings were made more precise to improve the user's understanding.