Add Defiedge platform assets and link
Added SVG icons for the new platform "Defiedge". The public and source assets directories have the new Defiedge.svg file, essential for visual identification of such platform. Also, a new URL link for the Defiedge platform is added in the "pool.js" file. This update enables the integration of Defiedge platform for better user interaction and easy access to the platform.