Update SwipeDetalization and odos-swap files for network changes
This update performs enhancement on SwipeDetalization and odos-swap for changes in network. Now the program will clear the data and load payout information whenever there is a change in network. It has also adjusted the tokensToBalanceUpdate function in odos-swap to accommodate stablecoinSecondTokens. Additionally, when the networkId changes, the balance and quotaResponseInfo are reset, and contract and data are reinitialized. Console logs have been updated to reflect these changes as well. This update should improve the app's responsiveness and accuracy to network changes.