Filter out zero-proportion outputTokens in ZapForm
The console.log was checking the proportion for outputTokens in ZapForm component. To skip the tokens which had 0 proportions while processing, a filter was added to clean the outputTokens array. This change prevents unnecessary processing of 0 proportion tokens improving efficiency and performance. Additional console logs were also added for better debugging and tracking of requestOutputTokens and requestInputTokens.